Refactoring components in the e2e project
