+++ e2e/src/lib.rs
@@ -1,1 +1,9 @@
+ethcontract::contract!("dex-contracts/build/contracts/BatchExchange.json");
+ethcontract::contract!("dex-contracts/build/contracts/IERC20.json");
+ethcontract::contract!("dex-contracts/build/contracts/IdToAddressBiMap.json");
+ethcontract::contract!("dex-contracts/build/contracts/IterableAppendOnlySet.json");
+ethcontract::contract!("dex-contracts/build/contracts/TokenOWL.json");
+ethcontract::contract!("dex-contracts/build/contracts/ERC20Mintable.json");
+
 pub mod common;
+pub mod stablex;
